requirements-test.txt
# Testing dependencies
pytest==7.4.3
pytest-xdist==3.8.0
pytest-asyncio==0.21.1
httpx==0.27.0
requests==2.31.0
pytest-dotenv==0.5.2

# Development dependencies
flake8==7.1.0
asyncpg==0.29.0
aio-pika==9.4.3
pydantic==2.8.2
structlog==24.4.0
jsonschema==4.22.0
